Fix for - Issue #3: Box.addDisplayArea() truncates area - Issue #4: Box edges overlap when drawn
Verified with the grin viewer running on Windows PC, jdk 1.6.0_01, with scaling turned off.  Both issue notes include a section of a grin script to verify the fix.  Note that without the scaling turned off (I have changed GenericMain's SCALE_DIVISOR from 2 to 1), the artifacts did not show up as they should.

Many thanks to Mark for the assistance in reproducing the problems.

git-svn-id: https://svn.java.net/svn/hdcookbook~svn@67 e5771d8f-0841-e040-ffdb-cc069419ffcf