Created a new animation framework.  This should make it much easier to make an xlet with a main animation loop.  See grin/doc-files/index.html under "Animation Framework" for an explanation.
The modifications to GrinView (in the file GenericMain, below)
got kind of complex, because GrinView has a more complex animation
loop than a typical xlet.  That's because GrinView has features
like scaling 1920x1080 down to an easier size to deal with, and
skipping ahead a number of frames.

The book menu was modified to use the new framework.

This framework enables a more optimized drawing algorithm.
See RenderArea and AnimationClient for details.  A subsequent
putback will modify the GRIN Show so that it takes advantage
of these optimizations.

Putback not reviewed -- but these changes will be reviewed
when the changes to the GRIN show are made.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@74 e5771d8f-0841-e040-ffdb-cc069419ffcf