Issue #6, Create binary show file format
Introducing a binary format for GRIN show.  The writer at grin/jdktools/binaryconverter/src/ directory reads in a text-based GRIN script and writes it out in a binary format to a filesystem. The reader at grin/library/src/com/hdcookbook/grin/io/binary reads in this binary format file.

MenuXlet's Director is changed to read in a binary file instead of menu.txt.  The build process for the MenuXlet uses the binary writer and generates a binary file "menu.grin" from menu.txt, and bundles that file in 00004.jar instead.

For the current syntax of the binary file, see the class comment of GrinBinaryWriter.java or GrinBinaryReader.java.  It's made of series of  data block based on DataOutputStream and DataInputStream, and it is subject to change.

Summary of changes are:

- Introducing com.hdcookbook.grin.io.binary for the binary file support.
- The classes at com.hdcookbook.grin.parser now moved to com.hdcookbook.grin.io.text.  ExtensionsParser got renamed to ExtensionsBuilder, and it is now under com.hdcookbook.grin.io.
- Manu getters were added to Feature subclasses, RCHandler subclasses, Command subclasses, Segment, and Show.  These are not expected to be used by xlets.  Getters have not been added.
- ExtensionsParser.parseCommand(...) is changed to ExtensionsBuilder.getCommand(...).
- Abstract RCHandler class now expects String in it's parameter, and so is CommandRCHandler.
- Introduced grin/samples/BinaryFormat, which does TumblingDuke with the binary file.
- jdktools/test directory got renamed to jdktools/grinviewer.

Tested by parsing ryan showfile, and making sure that hdcookbook disc image works properly on PC player after making MenuXlet read from a binary file.  

Thanks to Bill for numbers of explanation, suggestions and a through code review.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@94 e5771d8f-0841-e040-ffdb-cc069419ffcf