A few changes to the BDSigner to reflect Panasonic's comments. - Both the leaf and the root certificate require extensions mandated by MHP 12.5.9. - Certificate validity needs to be in GeneralizedTime format, MHP 12.5.5, - Subject line can contain UTF8String only, MHP 12.5.6.
We're generating the original self-signed certificates using KeyTools and using bouncycastle to generate a certificate in response to the certificate request file.  KeyTool tool doesn't provide any ways to add extensions, so I am overwriting jdk's CertAndKeyGen.java to do this manually in the method.  For GeneralizedTime, bouncycastle doesn't seem to provide a way to switch this by default, but it does use GeneralizedTime for time before 1950 and after 2050, so I'm setting the certificate validity boundry to be outside of that range.   Finally, for the UTF8String, the problem seemed to be about adding email address concept, so I'm using the newly added extensions instead of using subject line to pass in this info.

After this commit, the tool will generate two jar files, bdsigner.jar and bdsigner-bootclasspath.jar.  bdsigner-bootclasspath.jar needs to be prepended to the jdk bootclasspath at runtime.  I've only tested this with jdk 1.6.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@100 e5771d8f-0841-e040-ffdb-cc069419ffcf