Fixed a newly introduced bug in Translator where the draw area would sometimes be off by one frame.
Added a show for testing GRIN features, with an initial set
of segments for testing optimized draw.  This is put in a directory
where I hope we'll make an xlet version of the grinviewer program
in the fairly near future.

Added a case statement to run_book_menu so it can also be used to
run grinviewer on other stuff.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@121 e5771d8f-0841-e040-ffdb-cc069419ffcf