These are the changes made to the security classes:
1) Renaming of the security tools package from "net.java.bd.tools.bdsigner"
 ===> "net.java.bd.tools.security"

2) Added a BDCertGenerator tool that generates the root and application
certificates required for securing the BD-J applications environment.
A facility to customize some of the fields in the certificate is yet to come.

3) The BDSigner tool can now use the keystore generated by the
BDCertGenerator tool.

4) JarSigner no more supports -issuecert option. This code has been moved to
SecurityUtil class.  With this the only dependency on implementing JarSigner
is to add the BD-J specific manifest attribute to the signature file.
Will file an issue tracker request to work on this issue in near future
to re-generate the signature block file instead.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@172 e5771d8f-0841-e040-ffdb-cc069419ffcf