CHANGE 1:  (Reviewed by Sundar)
Re-factoring of "TranslatorModel" and "Timer" into a single
class called "InterpolatedModel".  An InterpolatedModel
is a feature that varies a set of parameters over time, doing
linear interpolation between keyframes.  This is exactly what
TranslatorModel was.  Timer is just a degenerate case of
InterpolatedModel - it has two keyframes (begin and end),
and zero parameters that are interpolated.

Doing this re-factoring sets us up for a third use of
InterpolatedModel, addressing issue 38, "support scale_model
feature".  The thing that represents the scaling can be an
InterpolatedModel with three parameters:  scaling factor,
and x,y coordinate of center.  Note that x,y might not vary
over time; InterpolatedModel handles the case of a non-varying
parameter efficiently.

CHANGE 2:  (reviewed by Chihiro)

Fixed some bugs with GrinView.  Now it uses insets properly,
and a race condition was fixed when an initial segment is
specified on the command line.




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@193 e5771d8f-0841-e040-ffdb-cc069419ffcf