1.Added "-debug" in the arg for BDSigner -- hdcookbook build
2. Corrected the sample xlet Reader,Writer (used for credentials) 
that had cut-paste errors.

3.Added "-binding" option to BDCertGenerator to generate the Binding Unit Root
Certificate, which is used for BUMF signing.

4.The BDSigner is updated to accept either "a BUMF" or "jar files" for signing
(the decision of what signing mechanism to be used is based on the given file
extension)
The input BUMF file name is of format: <disc_id>.bumf
The output Signature file would be: <disc_id>.busf. 

Note: The signed BUMF file is not yet tested on the player.

5.SecurityUtil.java enhanced to incorporate generation of Binding Unit Root
Certificate and signing of BUMF.

6.Bug fixes to Keystore cleanup- After deleting the aliases, the keystore file
was not being updated. 

7.Updated BDCertGenerator to generate unique random serial version number
for the certificates.

8.Updated the Credential Generator with a signing mechanism that directly uses
JDK APIs.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@195 e5771d8f-0841-e040-ffdb-cc069419ffcf