Added scaling_model feature.  A scaling_model can be applied to any number of fixed_image, image_sequence and/or box features, and scale them horizontally and/or vertically, centered around a given point.
In the bookmenu xlet, this is demonstrated with the segment
S:StartBonusVideo, which has the feature F:UserInput.Success
active.  This takes the user input keyboard, rotates it 360 degrees
around a vertical axis, then expands and fades it.  Looks great on
a PS/3!

This sample unintentionally shows another characteristic of BD-J
because this somewhat aggressive animation is attempted while a
new playlist is being selected.  On many players, the player is
so busy doing the playlist selecdtion that the animation gets
skipped entirely.  In a real title, you'd want to do the
animation first, and only then select the playlist, but for this
academic exercise it's worthwhile to leave it as it is.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@196 e5771d8f-0841-e040-ffdb-cc069419ffcf