Fix issue 27, "Animation Frameword documentation; docs refactoring".
Introduced https://hdcookbook.dev.java.net/animator.html and
https://hdcookbook.dev.java.net/grin.html to give an overview
of the goals and function of each of these components.  Added whole
new pages of technical documentation for the
com.hdcookbook.grin.animator package.  Big part of this documentation
were taken from a document Chihiro wrote; I guess some might consider
this to be "plagiarism" :-)

No code was hurt in the making of this putback.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@207 e5771d8f-0841-e040-ffdb-cc069419ffcf