Issue #54 FontHelper for jdktools
Adding an impl of AssetFinder.getFontHelper() to binary converter's main method, which just creates a new java.awt.Font instance.  This font will look different from what's expected during the xlet runtime, but it will hold the correct fontName, style and size that are needed to be recorded in the show binary file.

Reviewed by Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@220 e5771d8f-0841-e040-ffdb-cc069419ffcf