Enhancement for issue 32:  Now, any image can have a diferent anchor point and scaline factor.  This allows you to save image memory by flipping an image rather than loading a seperate flipped version.  All scaling/flipping is done at runtime; authors should be aware that a scale factor other than 1 or -1 might be slow on many players.
An image's anchor point can be top, middle or bottom vertically,
and left, middle or right horizontally.

This new feature is shown with the "spinning yin yang" effect.  It
now uses eight images instead of sixteen, because the last eight images
can be obtained by flipping the first eight.  The syntax in the text
file looks like this:

    feature image_sequence F:SpinningYinYangCCW {
            ( left   250 top     850    scale -1000 -1000 mills )
            ( left   250 top     850    scale -1000 -1000 mills )
            ( middle 313 top     850    scale -1000 -1000 mills )
            ( right  375 top     850    scale -1000 -1000 mills )
            ( left   250 middle  913    scale -1000 -1000 mills )

Each image in an image_sequence (or in a fixed_image) can have its
anchor point specified (top-left, middle-right, or any of nine
combinations), and can have a horizontal and a vertical scale 
specified.  It's specified in mills (that is, thousandths) for
consistency with scaling_model.

Other changes:

Moved animation framework documentation to its own subdirectory,
to emphasize that the animation framework can be used by itself.

Made the "main" target so it doesn't build the sample xlets,
and fixed the building of sample xlets.

Fixed an assertion failure with VisualRCHandlerHelper.  As a side
effect, removed the attempted (but undocumented and non-working) 
feature of making the cell at <0,0> be the default state.  Instead,
the states are now added in the intuitive left-to-right then top-to-bottom
order; the first real state encountered will be the default.  Note that
this default is overridden anyway by the UI state in the common case of
a visual RC handler tied to an assembly.

Eliminated a few "import xxx.*" statements.

Changed xlet so that bonus video only starts after the bonus animation
finishes.




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@222 e5771d8f-0841-e040-ffdb-cc069419ffcf