Issue #66, Desire an xlet that creates a text window for dumping debug log.
Ported a logging functionality from Sun's BD conformance test suite; the set of classes here can be used save a log to a file, and also put up a UI to display the file content.   As a sample, modified two StorageAccessXlets to use this logging jar as a classpath extension.

Reviewed by Jaya.

git-svn-id: https://svn.java.net/svn/hdcookbook~svn@263 e5771d8f-0841-e040-ffdb-cc069419ffcf