Issue 73, Logger can do automatically display log messages.
Adding a new class XletLogger.java, which is a wrapper class that works with the underlying classes from the test disc team.  Like the Logger and Screen classes, this XletLogger provides a bunch of static methods for logging, in a manner that is more suitable for a generic logging we need.

- The log goes to System.out and to the XletLoggerDialog automatically, and optionally to a file if the file location is set by the user.
- The log doesn't get cleared unless the user presses "3" to reset.

Also updating test xlets that relies on this logging library.

Reviewed by Sundar.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@274 e5771d8f-0841-e040-ffdb-cc069419ffcf