Fix for issue 70 (reviewed by Chiro):  Now, the mosaic builder tries a bunch of different widths, and picks the best one.  This can result in substantially better mosaics.  For the default mosaic, it'll let the width and height go up to 4096, which should work for the biggest images.
More optimization of mosaics is possible, e.g. with the techniques
in "Optimal Rectangle Packing: New Results" by Richard E. Korf, 2004.
We leave this for a later time, as the current simple algorigthm
works pretty well, and can always be tweaked with mosaic hints.
For the future, I think it would make sense to use this heuristic
to determine a resonable starting point, than work out in both
directions (width and height) from there, using the algorithm
in Korf 2004 page 5 (toward the bottom of column 1, under
"Searching the Space of Rectangles").

Tweaked the animation framework documentation.

Added a top-level "grin" target, to build the grin library and tools.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@282 e5771d8f-0841-e040-ffdb-cc069419ffcf