Issue 79 (Need some automation of the bumf.xml generation).
Adding BumfGenerator tool to generate a bumf.xml listing every files in the BDMV folder passed in as a parameter.  The user of the BridgeheadXlet is expected to name this xml as "sample.xml", sign it with bridgehead's keystore, and zip them with the BDMV folder and upload it to the player.

Also added a call in the SampleXlet to revert back VFSUpdate when an enter key is pressed.

Reviewed by Jaya.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@302 e5771d8f-0841-e040-ffdb-cc069419ffcf