Fixed: A special case handling for issue #81 reported by the customer.
The fix applies to only "-original-only" option of BDSigner.

The fix to handle re-signing of jar without additional
files added to it once it was signed worked well, but
for a jar file that didn't have any directory entries.
The present fix makes sure that a jar file without
directory entries also gets re-signed properly.

This bug is a show stopper for a customer. Committing the early version
of the fix immediately. As a follow up more testing will be done.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@310 e5771d8f-0841-e040-ffdb-cc069419ffcf