(Adding one more file to the pervious commit.)
Issue 85, Add a generic VFSUpdate reverting to Bridgehead

Changing the bridgehead disc structure so that the bridgehead xlet is launched as a first playlist instead of the first title in the titles array.  The topmenu still launches the first title (00000), and this is expected to be overwritten by the bridgehead user.   The bridgehead xlet offers a menu option to revert back the VFS update and go back to the optical disc structure.

Also added a check in the disc image download section, to make sure that the new disc image will not overwrite bridgehead xlet's bdjo/jar, or the first playlist of the index.bdmv to be other than bridgehead.  The VFS update is not going to happen if the error is detected.

Reviewed by Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@315 e5771d8f-0841-e040-ffdb-cc069419ffcf