Issue 90:  Multiple image map files.  Now, the asset finder lets you specify a list of image map files, which it merges together into one map.
Issue 10:  Auto-return assemblies.  Now, visual_rc_handler has
a flag called start_selected.  It defaults false for backwards
compatibilty.  If set true, then when the handler is activated
it checks the state of the assembly.  If the assembly is in one
of the activated states, it is coerced into the corresponding
selected state.

Issue 35: Pre-commands.  A segment now takes a list of commands
called "on_entry", which are run when the segment is activated.
This is useful for things like setting the state of a remote
control handler to some initial state.

Issue 36: Reset features.  There's a new reset_feature command
that resets a feature to the state it's set to when newly
activated.

<no bug ID>:  In image_sequence, the "repeat" flag was being
ignored -- all image sequences just repeated forever.  Usually
an end_command causes the show to move to a different segment,
so this wasn't noticed before.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@330 e5771d8f-0841-e040-ffdb-cc069419ffcf