Issue 93: Making playlist files
The PlayList tool converts the existing xxxxx.mpls file to an xml format and back.  In BD-ROM, parts of video streams are assembled into "playlists", in a fairly simple binary file.  This tool accepts an XML description of a playlist, and generates the binary file needed by a BD player, and vise versa.

Note that the playlist sets marks and descriptions which needs to correspond to the CLIPINF and STREAM files on the same disc.  BD spec 3-1 3.3 BD-ROM data model section provides a good overview of Movie PlayList.  The playlist tool currently do not provide integrety check with clip information.

Also converted hdcookbook sample disc image creation process to generate playlist from the xml files.

Reviewed by Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@336 e5771d8f-0841-e040-ffdb-cc069419ffcf