Fix for issue 97, "Need sync_display command". Also, creation of classload time measurement xlet that can be adapted to do other interesting statistical performance measurements.
The old way of doing sync_display was to make a command that returned
true from deferPendingCommands().  This is pretty counter-intuitive; it's
more natural to just have a sync_display command, rather than building a
flag into every command object.  This putback corrects this design
error.  In the old model, Command.deferPendingCommands() had to be
overridden to return true to give the sync_display behavior.  This method
is now final, so if anyone has subclassed it, they'll quickly be
lead to the javadocs in Command that tell them what to do.

The implementation of sync_display and segment_done have been
folded into the GrinXHelper class at runtime.  This was done to
minimize the number of classes at runtime, in order to have the
lowest possible startup time.

Added a new benchmark application to test classloading time on
a variety of players.  Perliminary results suggest that collapsing
classes is justified, at least where it doesn't compromise the
design too much.

While looking at this, I realized that there was a call to
Show.runPendingCommands() in Feature, that ran when a feature was
set up.  This was correct when written, but it's a fragile and
unnecessary thing to do.  It's also broken now, because there's
a small but finite chance that it might cause the state of the
show to change between when addDisplayAreas and paintFrame
are called.  This has been fixed by removing the extra and unnecessary
call.

Reviewed by Chihiro.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@341 e5771d8f-0841-e040-ffdb-cc069419ffcf