Fix for issue 96:  Add and remove animation clients
Added a resetAnimationClients method to AnimationEngine.  The hard
part here is re-initializing the drawing optimization information
in RenderContext.  That information is kept in a fairly non-OO way
(using an old-style doubly linked list) in order to get runtime
efficiency and prevent garbage generation during animation.  That's
a good result, but it does mean that this kind of change requires
a close look.

Added a new "Playground" functional test xlet to test features like
this that don't fit naturally into the cookbook menu xlet.

Made colors more space-efficient in the binary file by writing
the RGBA value as a single int, instead of four.

This putback was earthquake-tested up to a magnitude of 5.4
on the Rictor scale.

Reviewed by Chihiro


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@350 e5771d8f-0841-e040-ffdb-cc069419ffcf