Ussue 102:  Add "sticky images"
This adds support to ManagedImage for "sticky images".   If you
want an image to stay in memory once loaded, set it to sticky
(by calling ManagedImage.makeSticky()).  Each call to makeSticky()
must eventually be balanced with a call to unmakeSticky().

Took advantage of the sticky image support in Show, by having a list
of images that are to be sticky.  They're made sticky when the show is
initialized, and "unmade" sticky when the show is destroyed.

Changed GrinView so that it destroys the show before calling
System.exit(0).  This way, I was able to check that unmakeSticky()
works (by running GrinView on menu.txt, navigating to S:Initialize
then S:NullSegment, then exiting GrinView).

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@358 e5771d8f-0841-e040-ffdb-cc069419ffcf