Issue 103:  "Better support for game animations."  Implemented Feature.cloneSubgraph() so that a game can take a subtree of the scene graph, and create multiple copies.  Modified Group so that you can dynamically add and remove cloned features from a group's visible set.  Re-wrote the Gun Bunny game as GrinBunny, a game that's nearly functionally identical, but that uses the GRIN scene graph for display, state management, resource management, and remote control management.
*** BACKWARDS COMPATIBILTY WARNING - SEE NEXT ISSUE  ****

(no issue #) Fixed timer.  Timer was always running one frame too long, 
which meant that the fastest repeating timer would fire every *two* frames.  
This putback fixes that, which also means that all timers will run correctly
now, which is one frame faster than they used to.  There's an outside
chance that this might cause off-by-one-frame bugs in some existing
shows.

(no issue #) Added a "-director" argument to GrinView.  This lets you 
extend GrinView with your own director.  This is nice for shows where 
the BD-dependent functionality can be segregated outside of the director, 
because it means you can more easily execute the application inside 
GrinView, on desktop JDK.  For example, GrinBunny puts all of the game 
logic inside the director, which lets you debug and play the game within
GrinView.

(no issue #) Fixed "grinview -scale" - it used to generate a null
pointer exception under certain circumstances.

(no issue #) Made "linear-relative" translation preferred, and deprecated
"linear".  Made it a parse error if a show mixes the two, rather than just
ignoring all but the last one.  Added documentation
to the show BNF explaining why non-relative positioning was a bad idea.

(no issue #) Changed the method InterpolatedModel.getCurrValue(int) to
getField(int), for symmetry with setField(int, int).  This is a change
to the public API, but in a fairly new method that has an obvious
mapping, so if anyone has javac blow up on them, it shouldn't take
long to find!

Issue 49:  "Grinview with null-name segment, selecting assembly parts."
Now you can double-click on an unnamed segment and have it work.  Also,
if you double-click on an assembly part, instead of doing an
activate_segment, it will do an activate_part on the assembly.  This
is really handy for debugging the contents of assemblies.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@370 e5771d8f-0841-e040-ffdb-cc069419ffcf