Issue 109:  "Want extension classes to work without java_command_class"
Added the ability to record extension class names in the .grin binary
show file, so that extension features and commands can work even when
no java_generated_class is specified.  If a java_generated_class is
specified, the more obfuscator-friendly switch statement in the generated
class is used.

Added the ability to specify the name of the .grin file that the
compiler generates, which makes it possible to generate a binary 
file with the ".grn" extension for compatibility with 8.3 filesystems.  
This necessitated the addition of the "-binary" flag to GrinView, so 
it knows which kind of file to read.

Added a test to Playground's second_show.txt of the above two changes.
Playground now includes an extension feature to draw an arc on the screen.
Added grinview targets "run-grinview2" and "run-grinview2-binary" to
test this extension, both when reading the text and the binary files.

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@373 e5771d8f-0841-e040-ffdb-cc069419ffcf