Issue 110:  "Show setup takes much too long with a big show"
Made some structural changes to feature setups so that features generate
the "I've finished setup" command less often, and made segment's processing
of these commands much faster.

***  BACKWARDS COMPATIBILITY NOTE ***

The above change made it necesary to change the return type of
the abstract method Feature.setSetupMode from void to int.  Any
extension features will need a small change to adapt to this new API.

Made it so that only the two features that need to implent
SetupClient do.  This resulted in the elimination of some
unnecessary implementations of doSomeSetup() that were never
being called anyway.

Fixed up GrinView so we can collect cleaner profiling data, by turning
off the GUI and allowing control from stdin.  Also added a wait command.
Together these changes allow a run of GrinView to be scripted.

Made ManagedImage be completely thread-safe, even if multiple threads
prepare the same unprepared image simultaneously.  This never happens
with Grin's Show, but it's worth getting right.  

In ManagedImage, split prepare(Component) into prepare(), isLoaded()
and load(Component).  Added extensive documentation to ManagedImage's
main class documentation to say how to manage image loading and unloading.
Made addReference(), removeReference() and isReferenced() package-private,
since it only makes sense to call them from ImageManager.


Issue 113:  "Show not robust when active features aren't set up"

Originally, it was intended that a show could be sent to any segment,
at any time.  If an active feature hadn't finished its setup, it was
supposed to have just not displayed, until setup was completed.  However, 
the logic to support this was a little scrambled, with Segment calling
paintFrame etc. on features that weren't activated.  That's been fixed
in this putback.  The policy for features in the active clause of a segment
that aren't set up when the segment is activated is that when all of the
setup is done for the segment, all such features are activated simultaneously.


No issue number:  added "-binary" to some of the grinview autotest targets 
that needed it.


Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@380 e5771d8f-0841-e040-ffdb-cc069419ffcf