Issue 114:  "Need a key_released handler for games"
Added one like the key_pressed handler, called "key_released".  Put
code in Playground to test it.

Added a feature to GrinView to suppress key release events, so you can
test how a show would behave on a device that doesn't generate them.
Added a debug popup menu to GrinBunny to suppress key release events
for the same reason.

The key-event-suppression menu is done using a second show file that's
overlayed on top of the game's show file.  This is something the
animation manager has supported for a couple of months, but this is
the first sample of a serious use of this feature -- the Playground
sample app also does this, but in a more contrived and less useful
way.  Laying one show on top of another is a quite powerful feature
for keeping things decoupled, so it's perfect here, where the debug
menu is totally separate from the game.

Fixed up GrinBunny's build.xml and GrinBunnyXlet so that they're much more
generic, and can be adapted to another game with no changes to build.xml,
and a couple of trivial changes to GrinBunnyXlet.

Reviewed by Chihiro.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@386 e5771d8f-0841-e040-ffdb-cc069419ffcf