Issue 117:  "Null images in image sequences are broken"
When scaling was added, the GRIN compiler didn't account for
null images - now, a null image has a null placement.  At runtime,
ImageSequence deals with a null image by simply not adding the
drawRecord.  The animation framework understands this to mean that
the thing that was drawn in the previous frame should be erased in
this one.

Added a test of image_sequence to Playground that includes
"+" (repeat last image) and "-" (blank, that is, no image).

Reviewed by Chihiro



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@387 e5771d8f-0841-e040-ffdb-cc069419ffcf