<No issue number>
Adding a HelloWorldXlet project under the xlet directory.
This is meant to be a HelloWorld example for BD-J, and a basic
template for creating a compatible BD image from scratch.
The xlet source code is what I used to make a netbeans plugin
from.

Updated the security tool's CertGenerator to accept organization id 
with "0x" or "0X" prefix.

Reviewed by Bill Foote and Joe Wurzelbacher.  No licensing involved.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@389 e5771d8f-0841-e040-ffdb-cc069419ffcf