Issue 116:  GrinView-like GameXlet
This is the first installment on this issue...  More to come.

Created a generic xlet for running games.  Now, to create a new game,
you can just copy a short build.xml and vars.properties file, set a
few property values, and go.  The generic xlet includes functions for
debugging; a subsequent putback will significantly expand on this.
This generic xlet works for a game that uses one GRIN scene graph.

Used the new generic xlet for GrinBunny.  GrinBunny is thus an example
of how to set up a game xlet.

Added an extension feature to GrinBunny (a pie chart to show remaining
time).  This tests the build of GenericGame to make sure it works with
extension features.

Added Director.notifyDestroyed().  

Replaced GrinBunny's mechanism of using a segment called S:Finished 
and waiting for the show to get there; now, it just uses notifyDestroyed().
The old mechanism worked OK for GrinBunny, but in general was 
susceptible to a race condition where the show goes to S:Finished but
then immediately goes to another segment (due to a queued command), and
starts loading assets again.

Still to do:

    *  Make the "deploy" version of the xlet

    *  Get the three embedded files out of generic_build.xml, and
       use ant's token-replacement thing instead.

    *  Add the new debugging features mentioned in issue 116

<no issue number>:  Fixed RenderContextBase so that it will
work when there are no animation clients.

Reviewed by Chihiro.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@391 e5771d8f-0841-e040-ffdb-cc069419ffcf