Issue 121:  Removing fade or src_over from render chain doesn't cause redraw
Created a new abstract method on feature, markDisplayAreasChanged().  This
must call DrawRecord.setChanged() for any feature.  Note that any
extension features that use the GRIN framework will be forced to implement
this new method, which is a good thing, because that's where the call
to DrawRecord.setChanged() is needed.

Modified the animation framework so that calls to DrawRecord.setChanged()
will be effective, even before DrawRecord.setArea() is called.

Changed all features that modify the drawing of a child node so that
they call Feature.markDisplayAreasChanged() when the parent feature
is deactivated.

Added a regression test for this issue to Playground.

Fixed a bug in Fade when the alpha value is modified programmatically.
Now it notices whether or not the programmatically-set value is
SRC or SRC_OVER.

Reviewed by Mark Johnson




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@392 e5771d8f-0841-e040-ffdb-cc069419ffcf