More on 116, "GrinView-like GameXlet"...
Changes to GenericGame:

Made the build work on windows, by removing an assumption that the
root of the filesystem is "/".

Added variables xlet.lib.zip and se.lib.zip so that a project can
specify libraries to be linked in on both SE and the xlet.  This lets
media control libraries, and stubs for media control libraries be linked
in as appropriate.

Changed the handling of the extension.parser variable so that it works on
Windows, too.  Now, games that don't have an extension.parser should
leave the variable undefined.

Fixed the binary converter's auto-generated class so that it doesn't
assume com.hdcookbook.grin.Show is imported.

Removed an unintentional dependency on games having a font called "Lisa".
Now if a fonts directory exists, it is copied verbatum to AUXDATA.  Also
fixed GameXlet so it only tries to create a font manager if there was
a fonts directory at build time.

Made it an error in binaryconverter if an extension parser can't
be created.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@394 e5771d8f-0841-e040-ffdb-cc069419ffcf