Finishing up issue 116, "GrinView-like game xlet"
Added the ability to have xlet-specific code in the xlet_src
directory.  Used this in GunBunny for a skeleton PlayerWrangler
class to manage the player.  A future version of GrinBunny will
have a PlayerWrangler that shows the starfield video.

Added a debug log.  Added convenience method
Debug.printStackTrace(Throwable).  Added an overriding
definition of com.hdcookbook.grin.util.Debug to GenericGame
to redirect Debug.println() to the debug log.

Made a debug log listener that waits for telnet connections
on port 6000, and sends the debug log to (at most one) socket
when it gets a connection.

Added the deployment version of GenericXlet, which you can
get by compiling a game with "ant -Ddebug.or.deploy=deploy".
It turns off debugging and assertions, and doesn't have the
debug menu.

Fixed permissions file build of GameXlet, and added permissions for
ServerSocket.

Optimized GrinBunny drawwing by adding guarantee_fill features
to the saucer, the bunny, and each trooper.

Decided to not have debug menu items to navigate the show under test to
a certain segment or to change an assembly's state.  With a fully functional
xlet, that's not necessary, and so far it doesn't seem like it would be
very useful.



git-svn-id: https://svn.java.net/svn/hdcookbook~svn@396 e5771d8f-0841-e040-ffdb-cc069419ffcf