Updating GrinBunny to use FontStripText extension feature instead of the built-in Text feature for displaying score and remaining time of the game.
Messages for the game play - such as "Hit Enter to Play!" is still displayed using the standard Text.  The font used for FontStripText is currently "Arial", which is searched among the system font during the compile time.  This can be changed by modifying fontstrip-config.xml file under the src/com/hdcookbook/grinbunny dir.

Reviewed by Bill.


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@400 e5771d8f-0841-e040-ffdb-cc069419ffcf