(no issue #) - Narrowed assertion on InterpolatedModel.getField(int) to allow query of a programmatically-set value when the feature isn't activated.
(no issue #) - added a public static variable of GameXlet to give clients
a reference to the XletContext, if needed.

(no issue #) - changed the GrinBunny JAR file from 00000 to 12345,
and fixed a bug in GenericXlet that had hard-coded a JAR name of
00000 into the BDJO.

(no issue #) - Added post-spotless to GrinBunny build, to clean up the
generated fontstrip files.

Issue 123, "GenericGame and unsigned assets" - changed GenericGame so
that the assets are located in a directory on the optical disc,
rather than being bundled into the xlet JAR file.

Reviewed by Chihiro


git-svn-id: https://svn.java.net/svn/hdcookbook~svn@401 e5771d8f-0841-e040-ffdb-cc069419ffcf