Issue 127, "GenericXlet should assign the color keys with the HAVi API"
Added code in GenericGame to do that.   Also, added errata to
http://hdcookbook.com/errata/ in Chapter 19 that was discovered
when I tried to compile sample code 19.2 :-)

Added code to the debug log screen of GenericGame to report which
color key is pressed, to make it easy to verify correct key handling
on a variety of players.

Fixed bug with GenericGame's DebugLog, in the telnet server, where
it was not properly writing the last line of the log as soon as
it was available.




git-svn-id: https://svn.java.net/svn/hdcookbook~svn@404 e5771d8f-0841-e040-ffdb-cc069419ffcf